refactor: Operator(Symbol) enum 개선
