+++ src/test/java/study/basic/SetTest.java
@@ -1,3 +1,5 @@
+package study.basic;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
