+++ src/test/java/study/basic/StringTest.java
@@ -1,3 +1,5 @@
+package study.basic;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
